refact: convertMsToTimeString 모듈 분리 및 utils 폴더 생성
